pyproject.toml
[tool.poetry]
name = "juno"
version = "0.1.0"
description = "Discord Bot"
authors = ["ethanlchristensen <ethanlchristensen@outlook.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.13,<3.14"
discord-py = {git = "https://github.com/rapptz/discord.py.git"}
python-dotenv = "^1.1.0"
openai = ">=1.90.0,<2.0.0"
ollama = "^0.4.7"
yt-dlp = "^2025.9.23"
pynacl = "^1.5.0"
rich = "^14.0.0"
google-genai = "^1.20.0"
cryptography = "^45.0.4"
elevenlabs = "^2.4.0"
pytz = "^2025.2"
pillow = "^11.3.0"
anthropic = "^0.69.0"
ruff = "^0.13.3"
pyyaml = "^6.0.3"
tzdata = "^2025.2"
websocket-client = "^1.9.0"
numpy = "^2.3.3"
discord-ext-voice-recv = "^0.5.2a179"
pymongo = "^4.15.3"
fastapi = "^0.128.3"
motor = "^3.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 320
target-version = "py311"  # adjust to your Python version

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

